--- src/DellarteDellaGuerra.Infrastructure/Cannon/Mission/Battle/CannonTeamMissionLogic.cs
@@ -1,34 +1,0 @@
-﻿using System.Linq;
-using DellarteDellaGuerra.Infrastructure.Cannon.Mission.Siege.Spawn;
-using TaleWorlds.Core;
-using TaleWorlds.MountAndBlade;
-
-namespace DellarteDellaGuerra.Infrastructure.Cannon.Mission.Battle;
-
-public class CannonTeamMissionLogic : MissionLogic
-{
-    public override void OnAddTeam(TaleWorlds.MountAndBlade.Team team)
-    {
-        base.OnAddTeam(team);
-
-        TaleWorlds.MountAndBlade.Mission.Current.ActiveMissionObjects
-            .OfType<GenericCannon>()
-            .Where(script => script.Side.Equals(BattleSideEnum.Attacker) && team.IsAttacker)
-            .ToList()
-            .ForEach(script =>
-            {
-                script.Team = team;
-                script.SetForcedUse(true);
-            });
-
-        TaleWorlds.MountAndBlade.Mission.Current.ActiveMissionObjects
-            .OfType<GenericCannon>()
-            .Where(script => script.Side.Equals(BattleSideEnum.Defender) && team.IsDefender)
-            .ToList()
-            .ForEach(script =>
-            {
-                script.Team = team;
-                script.SetForcedUse(true);
-            });
-    }
-}
