--- src/DellarteDellaGuerra.Infrastructure/Cannon/Mission/Siege/Spawn/GenericCannonSpawner.cs
@@ -1,26 +1,0 @@
-using TaleWorlds.Core;
-using TaleWorlds.Engine;
-using TaleWorlds.MountAndBlade;
-using TaleWorlds.MountAndBlade.Objects.Siege;
-
-namespace DellarteDellaGuerra.Infrastructure.Cannon.Mission.Siege.Spawn;
-
-public enum Team { Attacker, Defender }
-
-public class GenericCannonSpawner : SpawnerBase
-{
-    [EditorVisibleScriptComponentVariable(true)]
-    public Team Team = Team.Attacker;
-
-    protected override void OnPreInit()
-    {
-        base.OnPreInit();
-        _spawnerMissionHelper = new SpawnerEntityMissionHelper(this);
-    }
-
-    public override void AssignParameters(SpawnerEntityMissionHelper spawnerMissionHelper)
-    {
-        var cannon = spawnerMissionHelper.SpawnedEntity.GetFirstScriptInFamilyDescending<GenericCannon>();
-        cannon.SetSide(Team == Team.Attacker ? BattleSideEnum.Attacker : BattleSideEnum.Defender);
-    }
-}
