--- src/DellarteDellaGuerra.Infrastructure/Cannon/Mission/Siege/Spawn/MissionSiegeWeaponsControllerPatch.cs
@@ -1,33 +1,0 @@
-using System;
-using System.Reflection;
-using DellarteDellaGuerra.Infrastructure.SiegeEngines.Port;
-using Harmony.DependencyInjection.Patches;
-using HarmonyLib;
-using TaleWorlds.Core;
-using TaleWorlds.MountAndBlade.Missions;
-
-namespace DellarteDellaGuerra.Infrastructure.Cannon.Mission.Siege.Spawn;
-
-public class MissionSiegeWeaponsControllerPatch : IPatch
-{
-    private static ICannonRegistry _cannonRegistry;
-
-    public MissionSiegeWeaponsControllerPatch(ICannonRegistry cannonRegistry)
-    {
-        _cannonRegistry = cannonRegistry;
-    }
-
-    public MethodInfo? TargetMethod =>
-        AccessTools.Method(typeof(MissionSiegeWeaponsController), "GetSiegeWeaponBaseType");
-
-    public MethodInfo? PatchMethod =>
-        AccessTools.Method(typeof(MissionSiegeWeaponsControllerPatch), nameof(Postfix));
-
-    public PatchType PatchType => PatchType.Postfix;
-
-    private static void Postfix(SiegeEngineType siegeWeaponType, ref Type __result)
-    {
-        if (_cannonRegistry?.GetCannon(siegeWeaponType.StringId) != null)
-            __result = typeof(GenericCannon);
-    }
-}
