+++ src/DellarteDellaGuerra.Infrastructure/DellarteDellaGuerra.Infrastructure.csproj
@@ -7,12 +7,6 @@
         <Platforms>x64</Platforms>
         <LangVersion>12.0</LangVersion>
         <Nullable>enable</Nullable>
-        <BuildForWindows>true</BuildForWindows>
-        <BuildForWindowsStore>true</BuildForWindowsStore>
-        <ModuleId>DellarteDellaGuerra.Core</ModuleId>
-        <ModuleName>Dell'Arte Della Guerra Core</ModuleName>
-
-        <GameFolder>D:\SteamLibrary\steamapps\common\Mount &amp; Blade II Bannerlord-v1.2</GameFolder>
     </PropertyGroup>
 
     <!-- Project properties -->
@@ -28,8 +22,6 @@
     </ItemGroup>
 
     <ItemGroup>
-        <PackageReference Include="Bannerlord.BuildResources" Version="1.1.0.104" PrivateAssets="all" IncludeAssets="runtime; build; native; contentfiles; analyzers; buildtransitive" />
-
         <PackageReference Include="IsExternalInit" Version="1.0.3" PrivateAssets="all" IncludeAssets="runtime; build; native; contentfiles; analyzers; buildtransitive"/>
         <PackageReference Include="SafeFluentXPath.Implementation" Version="1.0.0" />
         <PackageReference Include="NLog" Version="5.3.4"/>
@@ -59,11 +51,6 @@
       <Compile Remove="Cannon\PartyVisualTickSettlementVisualPatch.cs" />
     </ItemGroup>
 
-    <ItemGroup>
-      <Folder Include="Cannon\Infra\" />
-      <Folder Include="Cannon\Mission\Siege\" />
-      <Folder Include="Cannon\Util\" />
-    </ItemGroup>
 
     <Target Name="Copy technical config files to bin folder" BeforeTargets="BeforeBuild">
         <Message Importance="high" Text="Copy config files to bin folder" />
