--- src/DellarteDellaGuerra.Infrastructure/SiegeEngines/GenericCannonFactory.cs
@@ -1,26 +1,0 @@
-using System;
-using DellarteDellaGuerra.Infrastructure.Cannon.Mission.Siege.Spawn;
-using DellarteDellaGuerra.Infrastructure.SiegeEngines.Port;
-using TaleWorlds.MountAndBlade;
-
-namespace DellarteDellaGuerra.Infrastructure.SiegeEngines;
-
-public class GenericCannonFactory : ICannonFactory
-{
-    private readonly string _cannonId;
-
-    public GenericCannonFactory(string cannonId)
-    {
-        _cannonId = cannonId;
-    }
-
-    public Type CannonScriptType => typeof(GenericCannon);
-
-    public SpawnableArtilleryRangedSiegeWeapon CreateCannon() => new GenericCannon();
-
-    public void ConfigureSpawner(SpawnerEntityMissionHelper helper)
-    {
-        var cannon = helper.SpawnedEntity.GetFirstScriptInFamilyDescending<GenericCannon>();
-        // Configuration logic
-    }
-}
