+++ src/DellarteDellaGuerra.Infrastructure/SiegeEngines/Port/ICannonFactory.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using DellarteDellaGuerra.Infrastructure.Cannon.Mission.Siege.Spawn;
+using System;
 using TaleWorlds.MountAndBlade;
 
 namespace DellarteDellaGuerra.Infrastructure.SiegeEngines.Port;
