--- src/DellarteDellaGuerra.Infrastructure/SiegeEngines/Port/ICannonPrefabProvider.cs
@@ -1,12 +1,0 @@
-using TaleWorlds.Core;
-
-namespace DellarteDellaGuerra.Infrastructure.SiegeEngines.Port;
-
-public interface ICannonPrefabProvider
-{
-    string GetCampaignMapPrefabName(string cannonId, int wallLevel, BattleSideEnum side);
-    string GetCampaignMapProjectilePrefabName(string cannonId);
-    string GetCampaignMapReloadAnimationName(string cannonId);
-    string GetCampaignMapFireAnimationName(string cannonId);
-    int GetCampaignMapProjectileBoneIndex(string cannonId);
-}
