+++ src/DellarteDellaGuerra.Integration/Properties/launchSettings.json
