+++ src/DellarteDellaGuerra.Integration/SiegeEngines/Campaign/UI/CampaignMapSiegeEngineDeploymentIconEnricher.cs
@@ -1,0 +1,21 @@
+using DellarteDellaGuerra.Integration.SiegeEngines.Util.UI;
+
+namespace DellarteDellaGuerra.Integration.SiegeEngines.Campaign.UI;
+
+public class CampaignMapSiegeEngineDeploymentIconEnricher
+{
+    private const string SiegeEngineDeploymentIconBrushName = "CustomBattle.Siege.MachineIcon";
+
+    private readonly BrushStyleExtender _brushStyleExtender;
+
+    public CampaignMapSiegeEngineDeploymentIconEnricher(BrushStyleExtender brushStyleExtender)
+    {
+        _brushStyleExtender = brushStyleExtender;
+    }
+
+    public void AddCampaignMapSiegeEngineDeploymentIcon(string siegeEngineName, string siegeDeploymentIconSpritePath)
+    {
+        _brushStyleExtender.AddBrushStyle(siegeEngineName, siegeDeploymentIconSpritePath,
+            SiegeEngineDeploymentIconBrushName);
+    }
+}
