+++ src/DellarteDellaGuerra.Integration/SiegeEngines/CannonAvailabilityProvider.cs
@@ -1,0 +1,25 @@
+using System.Collections.Generic;
+using System.Linq;
+using DellarteDellaGuerra.Infrastructure.SiegeEngines.Port;
+using TaleWorlds.CampaignSystem.Party;
+using TaleWorlds.Core;
+using TaleWorlds.ObjectSystem;
+
+namespace DellarteDellaGuerra.Integration.SiegeEngines;
+
+public class CannonAvailabilityProvider
+{
+    private readonly ICannonRegistry _cannonRegistry;
+    private readonly MBObjectManager _mbObjectManager;
+
+    public CannonAvailabilityProvider(ICannonRegistry cannonRegistry, MBObjectManager mbObjectManager)
+    {
+        _cannonRegistry = cannonRegistry;
+        _mbObjectManager = mbObjectManager;
+    }
+
+    public IEnumerable<SiegeEngineType> GetAvailableCannons(PartyBase party, BattleSideEnum side) =>
+        _cannonRegistry.GetAllCannons()
+            .Select(cannon => _mbObjectManager.GetObject<SiegeEngineType>(cannon.Id))
+            .Where(se => se != null);
+}
