+++ src/DellarteDellaGuerra.Integration/SiegeEngines/Mission/Siege/Spawn/GenericCannon.cs
@@ -1,0 +1,17 @@
+using DellarteDellaGuerra.Infrastructure.SiegeEngines;
+using TaleWorlds.Core;
+using TaleWorlds.Engine;
+using TaleWorlds.ObjectSystem;
+
+namespace DellarteDellaGuerra.Integration.SiegeEngines.Mission.Siege.Spawn;
+
+public class GenericCannon : SpawnableArtilleryRangedSiegeWeapon
+{
+    [EditorVisibleScriptComponentVariable(true)]
+    public string SiegeEngineId = "";
+
+    public override SiegeEngineType GetSiegeEngineType()
+    {
+        return MBObjectManager.Instance.GetObject<SiegeEngineType>(SiegeEngineId);
+    }
+}
