+++ src/DellarteDellaGuerra.Integration/_Module/SubModule.xml
@@ -1,0 +1,38 @@
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<!-- This is an auto-generated SubModule.xml file -->
+<Module>
+  <Id value="$moduleid$" />
+  <Name value="$modulename$" />
+  <Version value="e$version$" />
+  <DefaultModule value="false" />
+  <ModuleCategory value="Singleplayer"/>
+  <ModuleType value="Community" />
+  <!-- Community Dependency Metadata -->
+  <DependedModules>
+    <DependedModule Id="Bannerlord.Harmony" DependentVersion="v$harmonyversion$"/>
+    <DependedModule Id="Native" DependentVersion="$gameversion$"/>
+    <DependedModule Id="SandBoxCore" DependentVersion="$gameversion$"/>
+    <DependedModule Id="Sandbox" DependentVersion="$gameversion$"/>
+    <DependedModule Id="CustomBattle" DependentVersion="$gameversion$"/>
+    <DependedModule Id="StoryMode" DependentVersion="$gameversion$"/>
+  </DependedModules>
+  <!-- Community Dependency Metadata -->
+  <!-- https://github.com/BUTR/Bannerlord.BLSE#community-dependency-metadata -->
+  <DependedModuleMetadatas>
+    <DependedModuleMetadata id="Bannerlord.Harmony" order="LoadBeforeThis" version="v$harmonyversion$"/>
+    <DependedModuleMetadata id="Native" order="LoadBeforeThis" version="$gameversion$.*"/>
+    <DependedModuleMetadata id="SandBoxCore" order="LoadBeforeThis" version="$gameversion$.*"/>
+    <DependedModuleMetadata id="Sandbox" order="LoadBeforeThis" version="$gameversion$.*"/>
+    <DependedModuleMetadata id="StoryMode" order="LoadBeforeThis" version="$gameversion$.*"/>
+    <DependedModuleMetadata id="CustomBattle" order="LoadBeforeThis" version="$gameversion$.*"/>
+  </DependedModuleMetadatas>
+  <SubModules>
+    <SubModule>
+      <Name value="DellarteDellaGuerra"/>
+      <DLLName value="DellarteDellaGuerra.Integration.dll"/>
+      <SubModuleClassType value="DellarteDellaGuerra.Integration.SubModule"/>
+      <Tags />
+    </SubModule>
+  </SubModules>
+  <Xmls/>
+</Module>
