+++ src/DellarteDellaGuerra.sln
@@ -39,6 +39,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "workflows", "workflows", "{
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DellarteDellaGuerra.Domain.Tests", "DellarteDellaGuerra.Domain.Tests\DellarteDellaGuerra.Domain.Tests.csproj", "{CE2969F6-44EB-4908-B2F1-FF882A6EFBE4}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DellarteDellaGuerra.Integration", "DellarteDellaGuerra.Integration\DellarteDellaGuerra.Integration.csproj", "{6F9D88D7-21D9-46A4-9A67-F4D3D66AE826}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
@@ -66,10 +68,14 @@ Global
 		{423C424E-22F0-464D-97AB-69E0E00D5F51}.Release|x64.Build.0 = Release|x64
 		
 		{CE2969F6-44EB-4908-B2F1-FF882A6EFBE4}.Debug|x64.ActiveCfg = Debug|x64
-        {CE2969F6-44EB-4908-B2F1-FF882A6EFBE4}.Debug|x64.Build.0 = Debug|x64
-        {CE2969F6-44EB-4908-B2F1-FF882A6EFBE4}.Release|x64.ActiveCfg = Release|x64
-        {CE2969F6-44EB-4908-B2F1-FF882A6EFBE4}.Release|x64.Build.0 = Release|x64
+								{CE2969F6-44EB-4908-B2F1-FF882A6EFBE4}.Debug|x64.Build.0 = Debug|x64
+								{CE2969F6-44EB-4908-B2F1-FF882A6EFBE4}.Release|x64.ActiveCfg = Release|x64
+								{CE2969F6-44EB-4908-B2F1-FF882A6EFBE4}.Release|x64.Build.0 = Release|x64
 		
+		{6F9D88D7-21D9-46A4-9A67-F4D3D66AE826}.Debug|x64.ActiveCfg = Debug|x64
+		{6F9D88D7-21D9-46A4-9A67-F4D3D66AE826}.Debug|x64.Build.0 = Debug|x64
+		{6F9D88D7-21D9-46A4-9A67-F4D3D66AE826}.Release|x64.ActiveCfg = Release|x64
+		{6F9D88D7-21D9-46A4-9A67-F4D3D66AE826}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
